Add restrictions on changing vehicle in permit
Refs: PV-680
